+++ frontend/src/metabase/common/components/DatePicker/SpecificDatePicker/SingleDatePicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./SingleDatePicker";
